+++ HealthCheck/HealthCheck/src/app/health-check/health-check.component.ts
@@ -2,29 +2,29 @@ import { Component, OnInit } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { environment } from './../../environments/environment';
 
+import { Observable } from 'rxjs';
+import { HealthCheckService, Result } from './health-check.service';
+
 @Component({
   selector: 'app-health-check',
   templateUrl: './health-check.component.html',
   styleUrls: ['./health-check.component.scss']
 })
 export class HealthCheckComponent implements OnInit {
-  public result_data?: Result;
+  public result: Observable<Result | null>;
 
-  constructor(private http: HttpClient) { }
+  constructor(public service: HealthCheckService) {
+    this.result = this.service.result;
+  }
 
   ngOnInit() {
-    this.http
-      .get<Result>(environment.baseUrl + 'api/health')
-      .subscribe(result => {
-        this.result_data = result;
-      }, error => console.error(error));
+    this.service.startConnection();
+    this.service.addDataListeners();
   }
-}
 
-interface Result {
-  checks: Check[];
-  totalStatus: string;
-  totalResponseTime: number;
+  onRefresh() {
+    this.service.sendClientUpdate();
+  }
 }
 
 interface Check {