+++ HealthCheck/HealthCheckAPI/Program.cs
@@ -1,5 +1,7 @@
+using HealthCheckAPI;
 using HealthCheckAPI.HealthCheck;
 using Microsoft.AspNetCore.Cors;
+using Microsoft.AspNetCore.SignalR;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -15,11 +17,15 @@ builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
+// Add a Cors Policy
 builder.Services.AddCors(options => options.AddPolicy(name: "AngularPolicy", cfg => {
   cfg.AllowAnyHeader();
   cfg.AllowAnyMethod();
   cfg.WithOrigins(builder.Configuration["AllowedCORS"]);
 }));
+
+// Add SignalR service
+builder.Services.AddSignalR();
 
 var app = builder.Build();
 
@@ -37,11 +43,25 @@ app.UseHttpsRedirection();
 
 app.UseAuthorization();
 
+// apply the CORS policy
+app.UseCors("AngularPolicy");
+
 app.UseHealthChecks(new PathString("/api/health"), new CustomHealthCheckOptions());
 
 app.MapControllers();
 
+// Offline Connection Test  
 app.MapMethods("/api/heartbeat", new[] { "HEAD" }, () => Results.Ok());
+
+// SignlR Hub
+app.MapHub<HealthCheckHub>("/api/health-hub");
+// SignlR broadcast message via minimal api instead of having a Controller
+app.MapGet("/api/broadcast/update2", async (IHubContext<HealthCheckHub> hub) =>
+{
+  await hub.Clients.All.SendAsync("Update", "test");
+  return Results.Text("Update2 message sent.");
+});
+
 
 app.MapFallbackToFile("/index.html");